--- remove_emojis.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-"""remove_emojis.ipynb
-
-Automatically generated by Colaboratory.
-
-Original file is located at
-    https://colab.research.google.com/drive/16Ku-UETT3N-WKxyVRWE7eLZGpHxsMBpv
-"""
-
-!pip install --upgrade emoji
-
-import pandas as pd
-import emoji
-
-def remove_emojis(text):
-    return emoji.demojize(text)
-
-# Load CSV data into a DataFrame
-df = pd.read_csv('measuring_hate_speech.csv')
-
-# Apply the remove_emojis function to the relevant column(s)
-df['text'] = df['text'].apply(remove_emojis)
-
-print(df['text'])
